\name{glm.coef}
\alias{glm.coef}
\title{GLM Coefficients}
\description{Extracts and formats coefficients from GLM models.}
\usage{
glm.coef(model, type = c("lm", "logistic"), digits = 3)
}
\arguments{
  \item{model}{A model object.}
  \item{type}{Type of model.}
  \item{digits}{Number of rounding digits.}
}
