\name{graph.catvar}
\alias{graph.catvar}
\title{Graph Categorical Variables}
\description{Internal function for plotting categorical variables.}
\usage{
graph.catvar(data = NULL, varname = NULL, out.label = NULL, 
	include.missing = c("no", "yes"), page.rows = NULL, ...)
}
\arguments{
  \item{data}{A dataset.}
  \item{varname}{Variable to plot.}
  \item{out.label}{Label for output.}
  \item{include.missing}{Whether to plot NAs.}
  \item{page.rows}{Rows per page.}
  \item{\dots}{Additional arguments.}
}
