\name{t1.by}
\alias{t1.by}
\title{
Creating Baseline Descriptive Tables and Exploring the Data
}
\description{
This function is an extension of t1 function, it will allow users to describe categorical and numerical variables in a dataframe by a grouping factor. It will also provide a p value for significance of differences in the distributions across groups. 
A list with 3 matrices is produced: Continuous.Variables, Categorical.Variables and Summary. The first 2 describe variables in detail and are good for exploring the data as well as finding outliers. 
The Summary matrix is ready to become Table 1 with descriptive or baseline data of groups being reported.
}
\usage{
t1.by(data = NA, selected.vars = NULL, group.var = NA, 
      miss.group = FALSE, var.names = NA, 
      table.na = c("ifany", "no", "always"), 
      cont.var = c("mean", "median", "auto", "manual"), 
      manual.mean = NA, cont.digits = 1, categ.test = c("chi2", "fisher"), 
      auto.convert = TRUE, max.levels = 4, export.table = NA, sep = ",")
}
\arguments{
  \item{data}{
an object of class data.frame containing the variables you wish to describe.
}
  \item{selected.vars}{
the variables in the dataframe you want to describe. These may be given without quotes (i.e. \code{selected.vars = c(age, sex)}), with quotes (i.e. \code{selected.vars = c("age","sex")}), or as a previously created character vector holding the names. Setting auto.convert to TRUE will automatically convert to factor string variables with less than 10 levels and exclude from the analysis string variables with 10 or more levels.
}
  \item{group.var}{
the grouping variable by which you want to describe your data. It may be given without quotes (i.e. \code{group.var = total_success}) or with quotes (i.e. \code{group.var = "total_success"}). This variable must be in the dataframe you specify for the data argument.
}
  \item{miss.group}{
if miss group is set to TRUE (default is FALSE) an extra column will be added with those that have missing data for the grouping variable.
}
  \item{var.names}{
variable names and levels are often not suitable for a publication table. This character vector allows you to specify variable names and levels manually, and should always be given in quotes, since these are display labels (which may contain spaces) rather than column names. It is recommended that you first run the function with this argument set to default = NA, then you can decide which names you want to change in the same order as printed in the Summary. You do not need to change all the names/ levels. The length of this vector does not need to be the same as the number of rows in the Summary. (i.e. A summary has the following rows: age, sex, sex_1male, sex_2female, race, race_1white, race_2asian, you only want to change the name for the 3rd and 4th rows then specify var.names = c(NA,NA,"Female","Male")) 
}
  \item{table.na}{
this argument applies to factor variables in the same way as for the table function in the base package. The default option ifany allows a more detailed description of the data. Although you will often set this option to no for the final publication table.
}
  \item{cont.var}{
summary measure for continuous variables. Defaults to "mean" = mean (sd). Other options are "median" = median (interquartile range), "auto" in which case a normal distribution test (shapiro wilk or lilliefors depending on sample size) will be applied to determine if the mean or median should be used variable-by-variable. Finally if you choose "manual", then you need to provide the manual.mean argument to specify which variables should be summarized by mean (sd) and which by median (IQR)
}
  \item{manual.mean}{
if cont.var is set as "manual" then this argument should be a logical vector having the same length as the number of continuous variables selected. TRUE will result in mean (sd) and FALSE will result in median (IQR). The order in which this vector is used is the same as the order in which continuous variable appear in the dataframe (as well as the same order as displayed in Continuous.Variables matrix generated by this function). 
}
  \item{cont.digits}{
Number of decimal places for continuous variables
}
  \item{auto.convert}{
if set to TRUE (the default) logical variables will be transformed to factor variables, string variables with less than 10 levels will be transformed to factors (otherwise excluded from the data) and levels of factor variables with 0 counts will be excluded.
}
  \item{max.levels}{
Maximum number of levels that the grouping variable can have. Defaults to 4. If the number of actual levels in the grouping variable is above 4, an error will be returned.
}
  \item{export.table}{
if you provide a name, (i.e. export.table = "table1") instead of the default value NA, then a .csv file is exported with the Summary matrix to the current working directory. The date is appended to the file name to avoid inadvertently deleting already created .csv files.
}
  \item{sep}{
separator for the csv file, if want to export to a csv file easily read by excel and you live in latin-america and use microsoft office in spanish use sep = ";"
}
\item{categ.test}{Select "fisher" or "chi2" for categorical variables.
}
}
\note{
The variable-name arguments \code{selected.vars} and \code{group.var} may be
supplied without quotes, with quotes, or (for \code{selected.vars}) as a
previously created character vector. If you have an object in your workspace whose
name matches a column you intend to reference unquoted, the value of that object
is used; in that unusual case, quote the column name to avoid ambiguity. The
\code{var.names} argument is different: it holds display labels (which may contain
spaces and punctuation), so it must always be given in quotes.
}
\value{
A list with 3 matrices is generate by function t1: Continuous.Variables, Categorical.Variables and Summary
  \item{Continuous.Variables}{Matrix with summary measures for all continuous variables including non.missing observations, mean, sd, median, p25, p75, min, max, p.value for shapiro.wilk test and lilliefors test}
  \item{Categorical.Variables}{Matrix with summary measures for categorical variables including non.missing observations, counts and percentages with one decimal place}
  \item{Summary}{Matrix with summary measures for categorical and continuous variables in the same order as they appear in the data.frame (or applied in the selected.vars argument, if this argument is specified). This matrix is presented in an abbreviated form as commonly displayed in publications}
}
\author{
Eugenio A. Maul MD MPH and Maria Angelica Paulos MD MPH
}
\seealso{
\code{\link{codebook}} To explore a dataframe numerically including missing values.

\code{\link{inspect.plot}} To display graphs with the distribution of all variables on the dataset making it easy to detect outliers and deviations from a nomrmal distribution in cotinuous variables.

\code{\link{t1}} To explore or summarize variables in a dataframe without a grouping factor.

\code{\link{t23}} To generate tables with coefficients of univariable and multivariable regression. To generate an appropriate object for plotting the regression results in a forest plot like fashion using the ramu.plot function.

\code{\link{ramu.plot}} To plot the regression result using a forest plot like format using an object created by the t23 function.

\code{\link{cluster.corr}} To calculate Pearson and Spearman Correlation Coefficient in data with Clustering.
}
\examples{
data(drops)

## 1 Simple Example (group.var quoted)
grouped.table.1 <- t1.by(data = drops, group.var="total_success")

## 2 The same with group.var unquoted
grouped.table.2 <- t1.by(data = drops, group.var=total_success)

## 3 More advanced example with unquoted variable selection
data(drops)
drops2 <- drops; drops2$total_success <- as.character(drops2$total_success) ; 
	x <- drops2$total_success;
drops2$total_success[x=="0"] <- "Success = No"; drops2$total_success[x=="1"] <- "Success = Yes"; 
drops2$total_success <- as.factor(drops2$total_success)  
grouped.table.3 <- t1.by(data = drops2 , group.var=total_success, 
	selected.vars = c(age, sex, va_log, eye, md), 
	var.names=c("Age (years)", "Sex", "Visual Acuity (LogMar)", "Eye", "Mean Deviation (dB)"), 
	table.na = "no")

## 4 Same as before, but export the table
data(drops)
drops2 <- drops; drops2$total_success <- as.character(drops2$total_success) ; 
x <- drops2$total_success;
drops2$total_success[x=="0"] <- "Success = No"; 
drops2$total_success[x=="1"] <- "Success = Yes"; 
drops2$total_success <- as.factor(drops2$total_success)  
grouped.table.4 <- t1.by(data = drops2 , group.var=total_success, 
	selected.vars = c(age, sex, va_log, eye, md), 
	var.names=c("Age (years)", "Sex", "Visual Acuity (LogMar)", "Eye", "Mean Deviation (dB)"), 
	table.na = "no", export.table = "Test_grouped_table")
}