\name{codebook}
\alias{codebook}
\title{
Exploring a Dataframe imported into R (Step 1)
}
\description{
Using this function is the first step to find out the structure and contents of a dataset imported into R.
A list is generated with:
1) Number of variables (columns) and number of observations (rows) in the dataset.
2) Types of variables and number of each type.
3) Number of observations (rows) with complete data. (no missing data)
4) Matrix with variable by variable info on:
	a) Variable name
	b) Variable type
	c) Number of missing values
	d) General Variable summary

This function is complemented by the following functions:
  - subset() to choose which variables and under which criteria to keep for the analysis.
}
\usage{
codebook(data = NA, blanks = c("separate", "missing"), 
         selected.vars = NULL, cont.digits = 1, export.table = NA, 
         sep = c(",", ";"))
}
\arguments{
  \item{data}{
an object of class data.frame imported into R that you want to explore before beginning analysis. No quotes needed.
}
  \item{blanks}{
blanks are empty, but non missing variables. For instance a character variable with an observation = "   ", will not be considered missing, even if converted to a factor.
Stating blanks = "separate" (the default) will add to the general variable summary a separate report to missing observations and blank observations.
Stating blanks = "missing" will count empty observations (those with only whitespace) as missing. But note that to actually convert them to missing (NA), you will need to do this separately.
}
  \item{selected.vars}{
a character vector specifying which variables in the dataframe you want to explore. This is typically left blank, because when exploring a dataframe imported you want a general view of everything.
}
  \item{cont.digits}{
Number of decimal places for continuous variables
}
  \item{export.table}{
if you provide a name, (i.e. export.table = "codebook") instead of the default value NA, then a .csv file is exported with the Summary matrix to the current working directory. The date is appended to the file name to avoid inadvertently deleting already created .csv files.
}
  \item{sep}{
separator for the csv file, if want to export to a csv file easily read by excel and you live in latin-america and use microsoft office in spanish use sep = ";", otherwise, use the default sep = ","
}
}
\value{
A list with 4 or 5 matrices is generated by function codebook:
  \item{dimensions}{Matrix with number of observations (rows) and variables (columns) in the dataset}
  \item{vartypes}{Matrix with number of different variable types in the dataframe. (i.e. numeric, factor (categorical), logical, character(free text))}
  \item{complete.data}{Matrix with: proportion of rows (observations) with full data}
  \item{summary}{Matrix with variable by variable information on:
  a) name 
  b) variable type
  c) proportion of cases with missing value for that observation
  d) median (interquartile range) for numeric variables. number of levels for categorical (factor) variables. 
  e) range (min-max values) for numeric variables, reference category (with its relative frequency within the variable) for categorical (factor) variables.}
  \item{exported files}{If option export.table is provided a name (i.e. export.table="codebook") then a .csv file is exported to the working directory with all the matrices generated above in the same worksheet.}
}

\author{
Eugenio A. Maul MD MPH, and Maria Angelica Paulos MD MPH
}

\seealso{
\code{\link{inspect.plot}} To display graphs with the distribution of all variables on the dataset making it easy to detect outliers and deviations from a nomrmal distribution in cotinuous variables.

\code{\link{t1}} To explore or summarize variables in a dataframe without a grouping factor.

\code{\link{t1.by}} To explore or summarize variables in a dataframe across a grouping variable.

\code{\link{t23}} To generate tables with coefficients of univariable and multivariable regression. To generate an appropriate object for plotting the regression results in a forest plot like fashion using the ramu.plot function.

\code{\link{ramu.plot}} To plot the regression result using a forest plot like format using an object created by the t23 function.

\code{\link{cluster.corr}} To calculate Pearson and Spearman Correlation Coefficient in data with Clustering.

}
\examples{
data(drops)
## 1 a simplified version of codebook function
codebook.example <- codebook(data = drops)

## 2 exporting the table to excel compatible file
## code not run
## codebook.example <- codebook(data = drops, export.table="codebook_test")
## will export a file called codebook_test.csv to the working directory

## 3 exporting the table to excel in spanish compatible file
## code not run
## codebook.example <- codebook(data = drops, export.table="codebook_test", sep=";")

## 4 exporting the table to excel in spanish compatible file only for some of the variables
## code not run
## codebook.example <- codebook(data = drops, export.table="codebook_test", sep=";", selected.vars=c("age","sex","va_log","iop_prev","md","diffic_level"))


}

