\name{graph.contvar}
\alias{graph.contvar}
\title{Graph Continuous Variables}
\description{Internal function for plotting continuous variables.}
\usage{
graph.contvar(data = NULL, varname = NULL, out.label = NULL, graph.range = NULL, run.independently = c("yes", "no"), page.rows = NULL, ...)
}
\arguments{
  \item{data}{A dataset.}
  \item{varname}{Variable to plot.}
  \item{out.label}{Label for output.}
  \item{graph.range}{Axis range.}
  \item{run.independently}{Standalone run flag.}
  \item{page.rows}{Rows per page.}
  \item{\dots}{Additional arguments.}
}
