\name{ramu.plot}
\alias{ramu.plot}
\title{
Plot Regression Results Using a Forest Plot Like Format
}
\description{
This functions requires an object created by the t23 function, which contains regression results. The ramu.plot function applied to such object will generate a plot that displays coefficients and their confidence intervals allowing for a very visual (rather than numeric) appreciation of associations of effects their magnitude and precision of the estimations.
}
\usage{
ramu.plot(data = NA, plot.pdf = FALSE, no.screen = FALSE, 
          plot.name = "figure1", plot.height = 9, plot.width = 8, 
          xlim.coord = NULL, x.ticks = NULL, xaxis.title = NULL, 
          h.coeff.align = NULL, v.coeff.align = -0.3, 
          coeff.labels = c("name.p", "name.coeff.p", "name", 
          "full.1.line", "full.2.line"), plot.order = c("increase", 
          "decrease", "signif.increase", "signif.decrease", "none"), 
          signif.col = "red", non.signif.col = "navyblue", 
          right.censor = NULL, left.censor = NULL, main.text = "", 
          axis.title.mag = 1.5, axis.label.mag = 1.5, axis.line.width = 1.5, 
          point.mag = 2, point.type = 18, ci.line.width = 1.5, 
          coeff.labels.mag = 1.25, vertical.line.width = 2, 
          plot.margins = c(8.1, 2.1, 4.1, 2.1), ...)
}
\arguments{
  \item{data}{
an object of list class generated by the t23 function.
}
  \item{x.ticks}{
essential in case of logarithmic modes. (Logistic, Cox, Poisson and negative binomial.)
numeric vector specifying where you want tick marks and labels added on the x axis. 
Minimize the number of ticks for easier readability of the graph. 
For logistic or gee.logistic specify in exponentiated units (i.e. Odds Ratio rather than log Odds difference. x.ticks = (0.25,0.5,1,2,4)). For Poisson and negative binomial specify the ticks as Incidence Rate Ratios (the same exponentiated logic applies).
}
  \item{xlim.coord}{
numeric vector with the minimum and maximum values in the X axis of the plot (i.e. xlim.coord=c(-10,20) will create a plot that goes from -10 to +20) The default value NULL will lead to a plot with a range for x axis that is twice the range determined by the minimum value of the lower bound for all confidence intervals (as plotted) and the maximum value of the upper bound for all confidence intervals (as plotted). If the variable labels are long, you might need to increase the size manually or decrease the text size with the coeff.labels.mag argument. The range for xlim.coord values should be broader than the range specified for x.ticks 
}
  \item{xaxis.title}{
a character vector to specify the title you want to provide for the x axis (i.e. "Odds Ratio for Glaucoma Progression (95\% CI)")
}
  \item{plot.pdf}{
if set to TRUE (default is FALSE) then a pdf file is generated to a file in the working directory in addition to displaying the graph on the screen. Note that depending on the plot.height and plot.width argument specification the appearance on screen could be quite different from the exported pdf file.
}
  \item{no.screen}{
if set to TRUE (default is FALSE) irrespective of the plot.pdf argument, a pdf file is exported without displaying the plot on the screen. This option will allow the code to run faster after you have fine-tuned the plot parameters.
}
  \item{plot.name}{
defaults for "figure1" and current date_time are appended to the filename to avoid inadvertently overwriting an already created .pdf file.
}
  \item{plot.height}{
height of the plot in inches. This argument applies only when exporting the plot to a pdf file (plot.pdf = TRUE) Increasing the parameter number for height and width will make the text, lines and symbols in the plot look smaller. Defaults to 9. 
If you have a large number of variables increase this number leaving the width constant. Reduce the height if there are very few variables.
}
  \item{plot.width}{
width of the plot in inches. Defaults to 8, see plot.height for more details.
}
  \item{h.coeff.align}{
will specify horizontal position for coefficient labels. The default option, NULL will automatically align the coefficient labels to the right of the highest value for the upper bond of all confidence intervals plotted. Option "right next" will put the coefficient names to the right of the right end of the confidence interval for each variable with left justification. Option "left.next" will do the opposite.
You can also manually specify a single number, in which case all coefficients labels will be put in the same place. (justification will be left for for positive numbers and right for negative numbers) 
Finally you can specify a vector of the same length as the number of coefficients, in which case each number of the vector will represent the location for the coefficient. In the latter case if the max number of the vector is positive all will be left justified, else if the minimum number of the vector is negative all will be right aligned.
}
  \item{v.coeff.align}{
defaults to -0.3, this means that the coefficient label is plotted slightly below the confidence interval line by default. As a reference the space between lines is 1. This argument helps fit the names in a narrower graph, particularly when the confidence intervals have a wide range. Values from -0.3 to 0.3 are recommended depending on preference. For very broad confidence interval see right.censor and left.censor arguments below.
}
  \item{coeff.labels}{
This argument defines the structure of the label to be added next to the confidence intervals for plotted variables. Default option "name.p" will paste the p value after the variable names. Option "name.coeff.p" will add the coefficient and p value to the variable names. In a similar way "name" will add just the variable name, "full.1.line" will add the coefficient, confidence interval and p value in 1 line to the variable names. "full.2.line" will do the same as above but in 2 lines
}
  \item{plot.order}{
Specifies the order by which variables should be plotted. Defaults to plot.order = "increase" which will result in coefficients plotted in ascending order. The opposite will happen if you specify "decrease". The "signif.increase" and "signif.decrease" will do the same but group the results into those that are significant first and non-significant in second place.
}
  \item{signif.col}{
color for significant results. Default is "red"
}
  \item{non.signif.col}{
color for non-significant results. Default is "navyblue"
}
  \item{right.censor}{
if a non-null numeric value is specified, all confidence intervals that extend above this value will be trimmed (censored) at this point.
}
  \item{left.censor}{
if a non-null numeric value is specified, all confidence intervals that extend below this value will be trimmed (censored) at this point.
}
  \item{main.text}{
Allows specification of a superior title for the plot using a character vector. Rarely used for this type of plot.
}
  \item{axis.title.mag}{
magnification factor for text of the x-axis title.
}
  \item{axis.label.mag}{
magnification factor for text of the x-axis labels. Defaults to 1.5. 
}
  \item{axis.line.width}{
magnification factor for width of the x-axis line. Defaults to 1.5.
}
  \item{point.mag}{
magnification factor for plotted symbols to represent coefficients. Defaults to 2.
}
  \item{point.type}{
type of symbol to represent the coefficient, default is a diamond (point.type = 18) For more options see help the pch argument of ?par.
}
  \item{ci.line.width}{
magnification factor for plotted lines of the confidence intervals. Defaults to 1.5.
}
  \item{coeff.labels.mag}{
magnification factor for text of variable labels. Defaults to 1.25.
}
  \item{vertical.line.width}{
magnification factor for width of the dashed vertical line separating positive from negative effects. Defaults to 2. 
}
  \item{plot.margins}{
number of lines of margin for the graph specified in the following order bottom, left, top, and right. Defaut value is c(8.1, 2.1, 4.1, 2.1) 
}
  \item{\dots}{
optional arguments
}
}
\author{
Eugenio A. Maul MD MPH and Maria Angelica Paulos MD MPH
}
\seealso{
\code{\link{codebook}} To explore a dataframe numerically including missing values.

\code{\link{inspect.plot}} To display graphs with the distribution of all variables on the dataset making it easy to detect outliers and deviations from a nomrmal distribution in cotinuous variables.

\code{\link{t1}} To explore or summarize variables in a dataframe without a grouping factor.

\code{\link{t1.by}} To explore or summarize variables in a dataframe across a grouping variable.

\code{\link{t23}} To generate tables with coefficients of univariable and multivariable regression. To generate an appropriate object for plotting the regression results in a forest plot like fashion using the ramu.plot function.

\code{\link{cluster.corr}} To calculate Pearson and Spearman Correlation Coefficient in data with Clustering.

}
\examples{

data(drops)
################################################################################
## Univariable Examples
################################################################################

# First create the object to be plotted using t23 function
table.2 <- t23(data = drops, delete.miss.rows=TRUE, mod.type="univar", 
	mod.func="gee", model.family="binomial", y.var="total_success", idvar="idn")

# 1 Plot a Simple Graph on screen (Univariable associations) with default arguments
ramu.plot(data = table.2 )

# 2 Export the plot in #1 to a pdf file (code not run)
# ramu.plot(data = table.2, plot.pdf = TRUE )

# 3 Export the plot in #1 to a pdf file without displaying it on screen (code not run)
# ramu.plot(data = table.2, no.screen = TRUE )

# 3 Cutomized Plot
# to change the names of the coefficients you must do so in the t23 object created a priori
table.2b <- t23(data = drops, delete.miss.rows = TRUE, mod.type = "univar", 
 mod.func = "gee", model.family = "binomial", 
	y.var = "total_success", idvar = "idn", 
	coeff.names =c("Age (per 1 year)", "Female sex (vs male)", 
	"Visual Acuity (per 10 lines less)", "IOP (per 1 mmHg)", 
	"MD (per 1dB less damage)","Number of drops/day (per 1 extra drop)", 
	"OS (vs OD)", "Second eye instilled (vs First)", 
	"Low-Mod Diffic to use drops (vs Low)", "Mod-High Diffic to use drops (vs Low)", 
	"High Diffic to use drops (vs Low)"))

# now run the function
ramu.plot(data = table.2b, xlim.coord = c(log(0.23),log(50)), x.ticks = c(0.25,0.5,1,2), 
	xaxis.title = "OR (95\% CI) \nSuccessful Eyedrop Instillation", 
	h.coeff.align = log(2), coeff.labels = "name",
	plot.order = "decrease", signif.col = "red", non.signif.col = "blue")

################################################################################
## Multivariable Examples
################################################################################

# 4 Cutomized Plot
# to change the names of the coefficients you must do so in the t23 object created a priori

table.3 <- t23(data = drops, delete.miss.rows = FALSE, 
	mod.type="multivar", mod.func="gee", model.family="binomial", 
	y.var="total_success", idvar="idn", 
	coeff.names =c("Age (per 1 year)", "Female sex (vs male)", 
	"Visual Acuity (per 10 lines less)", "IOP (per 1 mmHg)", 
	"MD (per 1dB less damage)","Number of drops/day (per 1 extra drop)", 
	"OS (vs OD)", "Second eye instilled (vs First)", 
	"Low-Mod Diffic to use drops (vs Low)", "Mod-High Diffic to use drops (vs Low)", 
	"High Diffic to use drops (vs Low)"))

# now run the function
ramu.plot(data = table.3, xlim.coord = c(log(0.23),log(200)), x.ticks = c(0.25,0.5,1,2), 
	xaxis.title = "OR (95\% CI) \nSuccessful Eyedrop Instillation", 
	h.coeff.align = log(2), v.coeff.align=0.2, 
	coeff.labels="name.p",plot.order="increase",signif.col="red",non.signif.col="blue", 
	coeff.labels.mag=0.95, plot.pdf = TRUE)


}