\name{t1}
\alias{t1}
\title{
Creating Baseline Descriptive Tables and Exploring the Data
}
\description{
This function will allow users to describe categorical and numerical variables in a dataframe. A list with 3 matrices is produced: Continuous.Variables, Categorical.Variables and Summary. The first 2 describe variables in detail and are good for exploring the data as well as finding outliers. 
The Summary matrix is ready to become Table 1 with descriptive or baseline data of a single group being reported.
}
\usage{
t1(data = NA, selected.vars = NULL, var.names = NA, 
   table.na = c("ifany", "no", "always"), 
   cont.var = c("mean", "median", "auto", "manual"), 
   manual.mean = NA, cont.digits = 1, auto.convert = TRUE, 
   export.table = NA, sep = ",")
}
\arguments{
  \item{data}{
an object of class data.frame containing the variables you wish to describe.
}
  \item{selected.vars}{
a character vector specifying which variables in the dataframe you want to describe. Setting the auto.convert to TRUE will automatically convert to factor string variables with less than 10 levels and exclude from the analysis string variables with 10 or more levels. (i.e. selected.vars = c("age","sex")) 
}
  \item{var.names}{
variable names and levels are often not suitable for a publication table. This character vector allows you to specify variable names and levels manually. It is recommended that you first run the function with this argument set to default = NA, then you can decide which names you want to change in the same order as printed in the Summary. You do not need to change all the names/ levels. The length of this vector does not need to be the same as the number of rows in the Summary. (i.e. A summary has the following rows: age, sex, sex_1male, sex_2female, race, race_1white, race_2asian, you only want to change the name for the 3rd and 4th rows then specify var.names = c(NA,NA,"Female","Male")) 
}
  \item{table.na}{
this argument applies to factor variables in the same way as for the table function in the base package. The default option ifany allows a more detailed description of the data. Although you will often set this option to no for the final publication table.
}
  \item{cont.var}{
summary measure for continuous variables. Defaults to "mean" = mean (sd). Other options are "median" = median (interquartile range), "auto" in which case a normal distribution test (shapiro wilk or lilliefors depending on sample size) will be applied to determine if the mean or median should be used variable-by-variable. Finally if you choose "manual", then you need to provide the manual.mean argument to specify which variables should be summarized by mean (sd) and which by median (IQR)
}
  \item{manual.mean}{
if cont.var is set as "manual" then this argument should be a logical vector having the same length as the number of continuous variables selected. TRUE will result in mean (sd) and FALSE will result in median (IQR). The order in which this vector is used is the same as the order in which continuous variable appear in the dataframe (as well as the same order as displayed in Continuous.Variables matrix generated by this function). 
}
  \item{cont.digits}{
Number of decimal places for continuous variables
}
  \item{auto.convert}{
if set to TRUE (the default) logical variables will be transformed to factor variables, string variables with less than 10 levels will be transformed to factors (otherwise excluded from the data) and levels of factor variables with 0 counts will be excluded.
}
  \item{export.table}{
if you provide a name, (i.e. export.table = "table1") instead of the default value NA, then a .csv file is exported with the Summary matrix to the current working directory. The date is appended to the file name to avoid inadvertently deleting already created .csv files.
}
  \item{sep}{
separator for the csv file, if want to export to a csv file easily read by excel and you live in latin-america and use microsoft office in spanish use sep = ";"
}
}
\value{
A list with 3 matrices is generate by function t1: Continuous.Variables, Categorical.Variables and Summary
  \item{Continuous.Variables}{Matrix with summary measures for all continuous variables including non.missing observations, mean, sd, median, p25, p75, min, max, p.value for shapiro.wilk test and lilliefors test}
  \item{Categorical.Variables}{Matrix with summary measures for categorical variables including non.missing observations, counts and percentages with one decimal place}
  \item{Summary}{Matrix with summary measures for categorical and continuous variables in the same order as they appear in the data.frame (or applied in the selected.vars argument, if this argument is specified). This matrix is presented in an abbreviated form as commonly displayed in publications}
}
\author{
Eugenio A. Maul MD MPH and Maria Angelica Paulos MD MPH
}
\seealso{
\code{\link{codebook}} To explore a dataframe numerically including missing values.

\code{\link{inspect.plot}} To display graphs with the distribution of all variables on the dataset making it easy to detect outliers and deviations from a nomrmal distribution in cotinuous variables.

\code{\link{t1.by}} To explore or summarize variables in a dataframe across a grouping variable.

\code{\link{t23}} To generate tables with coefficients of univariable and multivariable regression. To generate an appropriate object for plotting the regression results in a forest plot like fashion using the ramu.plot function.

\code{\link{ramu.plot}} To plot the regression result using a forest plot like format using an object created by the t23 function.

\code{\link{cluster.corr}} To calculate Pearson and Spearman Correlation Coefficient in data with Clustering.

}
\examples{
data(drops)
## 1 a simplified version of t1 function
table.1.1 <- t1(data = drops)
## table.1; - to see the function results in deatil, code not run

## 2 using more options
table.1.2 <- t1(data = drops,selected.vars = c("age","sex","va_log",
	"iop_prev","md","diffic_level"),
	var.names=c("Age (years)","Sex","Visual Acuity (logMar)", "Baseline IOP (mmHg)", 
	"Mean Deviation (dB)", "Difficulty for Eyedrop Use"),cont.var="auto")

## 3 using even more options
table.1.3 <- t1(data = drops,selected.vars = c("age","sex","va_log",
	"iop_prev","md","diffic_level"),
	var.names=c("Age (years)","Sex","Visual Acuity (logMar)", "Baseline IOP (mmHg)", 
	"Mean Deviation (dB)", "Difficulty for Eyedrop Use"),table.na="no", cont.var="manual",manual.mean=c(TRUE,NA,FALSE,TRUE,TRUE,NA))

## 4 exporting the table in #3 using the sep = ";" option suitable for reading in MS Excel in Chile 
##	code not run
##	table.1.3 <- t1(data = drops,selected.vars = c("age","sex","va_log","iop_prev","md","diffic_level"),
##			var.names=c("Age (years)","Sex","Visual Acuity (logMar)", "Baseline IOP (mmHg)", 
##			"Mean Deviation (dB)", "Difficulty for Eyedrop Use"),table.na="no", cont.var="manual",
##			manual.mean=c(T,NA,F,T,T,NA), export.table="table1_example", sep=";")

}

